feat(LazyLoader): Improve performance by removing context API, reducing DEFAULT_DEVICE_PIXEL_RATIOS count; using PLACEHOLDER_IMAGE, rather than low res version of passed src
