+++ src/components/LazyLoader/__tests__/index.spec.js
@@ -4,11 +4,12 @@ import {
   renderComponent,
   createImgWithSrcset,
   createImgWithSrc,
-  createDiv
+  createDiv,
+  PROPS
 } from "../__mocks__/utils.mocks";
 
 describe("LazyLoader", () => {
-  it("should load a low res, blurred version of the image passed prior to loading", () => {
+  it("should load a placeholder image passed prior to loading", () => {
     const component = renderComponent({}, renderer.create, {
       createNodeMock: createImgWithSrc
     });
@@ -16,7 +17,7 @@ describe("LazyLoader", () => {
     expect(component).toMatchSnapshot();
   });
 
-  it.skip("should load a high res version of the image passed when the imageRef supports srcset", () => {
+  it("should load a high res version of the image passed when the imageRef supports srcset", () => {
     const component = renderComponent({}, renderer.create, {
       createNodeMock: createImgWithSrcset
     });
@@ -27,7 +28,7 @@ describe("LazyLoader", () => {
     expect(instance.state.imageRef).toMatchSnapshot();
   });
 
-  it.skip("should load a high res version of the image passed when the imageRef supports src", () => {
+  it("should load a high res version of the image passed when the imageRef supports src", () => {
     const component = renderComponent({}, renderer.create, {
       createNodeMock: createImgWithSrc
     });
@@ -38,7 +39,7 @@ describe("LazyLoader", () => {
     expect(instance.state.imageRef).toMatchSnapshot();
   });
 
-  it.skip("should load a high res version of the image passed when the imageRef supports a backgroundImage style", () => {
+  it("should load a high res version of the image passed when the imageRef supports a backgroundImage style", () => {
     const component = renderComponent({}, renderer.create, {
       createNodeMock: createDiv
     });
@@ -49,7 +50,7 @@ describe("LazyLoader", () => {
     expect(instance.state.imageRef).toMatchSnapshot();
   });
 
-  it.skip("should invoke componentDidUpdate and set state to a low res, blurred version of the image when the src changes", () => {
+  it("should invoke componentDidUpdate and set state to a placeholder image when the src changes", () => {
     const component = renderComponent({}, renderer.create, {
       createNodeMock: createImgWithSrcset
     });
@@ -64,5 +65,38 @@ describe("LazyLoader", () => {
     );
 
     expect(component.toJSON()).toMatchSnapshot();
+  });
+
+  it("should not load a high res version of the image passed when the load function is invoked with a value of false", () => {
+    const component = renderComponent({}, renderer.create, {
+      createNodeMock: createImgWithSrcset
+    });
+
+    const instance = component.getInstance();
+    const onLoadSpy = jest.spyOn(instance, "onLoad");
+    instance.load(false);
+
+    expect(onLoadSpy).not.toHaveBeenCalled();
+    expect(component).toMatchSnapshot();
+  });
+
+  it("should not error when calling the onLoad method if no ref is available", () => {
+    const component = renderComponent({}, renderer.create);
+
+    const instance = component.getInstance();
+    expect(() => instance.onLoad()).not.toThrow();
+  });
+
+  it("should invoke componentDidUpdate and not setState when src equals prevSrc", () => {
+    const component = renderComponent({}, renderer.create, {
+      createNodeMock: createImgWithSrcset
+    });
+
+    const instance = component.getInstance();
+    const setStateSpy = jest.spyOn(instance, "setState");
+    instance.componentDidUpdate(PROPS);
+
+    expect(setStateSpy).not.toHaveBeenCalled();
+    expect(instance.state).toMatchSnapshot();
   });
 });
