+++ src/components/LazyLoader/constants.js
@@ -5,4 +5,4 @@ export const BLUR_STYLES = {
   transition: "80ms filter linear"
 };
 export const DEFAULT_TARGET_DENSITY = 1;
-export const DEFAULT_DEVICE_PIXEL_RATIOS = [1, 2, 3, 4, 5];
+export const DEFAULT_DEVICE_PIXEL_RATIOS = [1, 2];
