+++ src/index.js
@@ -1,4 +1,4 @@
-export { Button, Badge, RatingBadge, IconButton } from './components/Button';
+export { Button, Badge, RatingBadge, IconButton } from "./components/Button";
 export {
   DayTile,
   DayTileDate,
@@ -6,12 +6,12 @@ export {
   DayTileButton,
   DayTileMoreButton,
   DayTileOverflowButton
-} from './components/CalendarView';
-export { default as Card } from './components/Card';
-export { default as Column } from './components/Grid/Column';
-export { default as Container } from './components/Grid/Container';
-export { default as LoaderRing } from './components/LoaderRing';
-export { default as DrawerProvider } from './components/Drawer/Provider';
+} from "./components/CalendarView";
+export { default as Card } from "./components/Card";
+export { default as Column } from "./components/Grid/Column";
+export { default as Container } from "./components/Grid/Container";
+export { default as LoaderRing } from "./components/LoaderRing";
+export { default as DrawerProvider } from "./components/Drawer/Provider";
 export {
   default as Gradient,
   GradientStyles,
@@ -26,39 +26,39 @@ export {
   OVERLAY_STOPS_NOT_SPOTLIGHT,
   SPOTLIGHT_DEG,
   SPOTLIGHT_STOPS
-} from './components/Gradient';
-export { default as Header } from './components/Header';
-export { default as HeaderWithImage } from './components/Header/WithImage';
-export { default as Heading } from './components/Header/Heading';
-export { default as ImageCard } from './components/ImageCard';
-export { default as Banner } from './components/Banner';
-export { default as FeedbackInline } from './components/FeedbackInline';
+} from "./components/Gradient";
+export { default as Header } from "./components/Header";
+export { default as HeaderWithImage } from "./components/Header/WithImage";
+export { default as Heading } from "./components/Header/Heading";
+export { default as ImageCard } from "./components/ImageCard";
+export { default as Banner } from "./components/Banner";
+export { default as FeedbackInline } from "./components/FeedbackInline";
 export {
   ListContainer,
   ListRow,
   ListRowOverflow,
   Section,
   SectionItem
-} from './components/List';
-export { default as BottomSheet } from './components/BottomSheet';
-export { default as NavBar } from './components/NavBar';
-export { default as Row } from './components/Grid/Row';
-export { default as SearchInput } from './components/Input/Search';
-export { default as SearchComponent } from './components/SearchInput';
+} from "./components/List";
+export { default as BottomSheet } from "./components/BottomSheet";
+export { default as NavBar } from "./components/NavBar";
+export { default as Row } from "./components/Grid/Row";
+export { default as SearchInput } from "./components/Input/Search";
+export { default as SearchComponent } from "./components/SearchInput";
 export {
   default as SearchInputMobile
-} from './components/SearchInput/SearchInputMobile';
-export { SearchSuggest } from './components/SearchInput/Search.styles';
-export { default as Spacing } from './components/Spacing';
-export { default as Breadcrumb } from './components/Breadcrumbs';
-export { default as ContainerBlock } from './components/Container/Block.styles';
+} from "./components/SearchInput/SearchInputMobile";
+export { SearchSuggest } from "./components/SearchInput/Search.styles";
+export { default as Spacing } from "./components/Spacing";
+export { default as Breadcrumb } from "./components/Breadcrumbs";
+export { default as ContainerBlock } from "./components/Container/Block.styles";
 export {
   ResponsiveImage,
   StaticImage,
   ThumbnailCircle,
   StyledImageSeo
-} from './components/Image';
-export { Text } from './components/Text';
+} from "./components/Image";
+export { Text } from "./components/Text";
 export {
   AlertIcon,
   CAIcon,
@@ -147,33 +147,33 @@ export {
   PrintIcon,
   LaunchIcon,
   PlayCircularIcon
-} from './components/Icons';
+} from "./components/Icons";
 export {
   default as DeviceSizeProvider
-} from './components/DeviceSize/Provider';
+} from "./components/DeviceSize/Provider";
 export {
   Consumer as DeviceSizeConsumer
-} from './components/DeviceSize/Context';
-export { default as DisplayFor } from './components/DeviceSize';
+} from "./components/DeviceSize/Context";
+export { default as DisplayFor } from "./components/DeviceSize";
 export {
   default as ScrollObserverProvider
-} from './components/ScrollObserver/Provider';
+} from "./components/ScrollObserver/Provider";
 export {
   default as ScrollObserverConsumer
-} from './components/ScrollObserver/Consumer';
-export { default as Portal } from './components/Portal';
-export { default as colors } from './theme/colors';
-export { default as themes } from './theme/colorThemes';
-export { default as constants } from './theme/constants';
-export { default as spacing } from './theme/spacing';
-export { default as typography } from './theme/typography';
-export { default as zIndex } from './theme/zIndex';
+} from "./components/ScrollObserver/Consumer";
+export { default as Portal } from "./components/Portal";
+export { default as colors } from "./theme/colors";
+export { default as themes } from "./theme/colorThemes";
+export { default as constants } from "./theme/constants";
+export { default as spacing } from "./theme/spacing";
+export { default as typography } from "./theme/typography";
+export { default as zIndex } from "./theme/zIndex";
 export {
   largeAndUp,
   mediumAndUp,
   xLargeAndUp,
   smallAndUp
-} from './theme/mediaQueries';
+} from "./theme/mediaQueries";
 export {
   PrimaryText,
   SecondaryText,
@@ -181,7 +181,7 @@ export {
   Link,
   LinkCta,
   BoldText
-} from './components/Text';
+} from "./components/Text";
 
 export {
   Toggle,
@@ -194,44 +194,40 @@ export {
   Input,
   ButtonGroup,
   QtySelector
-} from './components/Input';
-export { LinkList, LinkListItem } from './components/Link';
-export { default as Modal } from './components/List/Modal';
-export { default as Backdrop } from './components/List/Backdrop';
-export { BackdropProvider } from './components/List/BackdropContext';
-export { ContainerProvider } from './components/List/Context';
+} from "./components/Input";
+export { LinkList, LinkListItem } from "./components/Link";
+export { default as Modal } from "./components/List/Modal";
+export { default as Backdrop } from "./components/List/Backdrop";
+export { BackdropProvider } from "./components/List/BackdropContext";
+export { ContainerProvider } from "./components/List/Context";
 
-export { default as AdaptiveModal } from './components/Modal';
+export { default as AdaptiveModal } from "./components/Modal";
 export {
   default as AdaptiveActionBar
-} from './components/Modal/DefaultActionBar';
+} from "./components/Modal/DefaultActionBar";
 export {
   MODAL_SIZE_SMALL,
   MODAL_SIZE_MEDIUM,
   MODAL_SIZE_LARGE,
   MODAL_SIZE_XLARGE
-} from './components/Modal/size';
-export { ModalConsumer, withModal } from './components/Modal/context';
-export { default as AdaptiveBackdrop } from './components/Backdrop';
+} from "./components/Modal/size";
+export { ModalConsumer, withModal } from "./components/Modal/context";
+export { default as AdaptiveBackdrop } from "./components/Backdrop";
 
-export { default as PopOver } from './components/PopOver';
-export { default as Tooltip } from './components/Tooltip';
+export { default as PopOver } from "./components/PopOver";
+export { default as Tooltip } from "./components/Tooltip";
 export {
   DeviceConnectionProvider,
   DeviceConnectionConsumer,
   DisplayOn
-} from './components/DeviceConnection';
-export { default as SeatTooltip } from './components/Tooltip/SeatTooltip';
-export { default as SeatData } from './components/Tooltip/SeatData';
-export {
-  LazyLoaderProvider,
-  LazyLoaderConsumer,
-  LazyLoader
-} from './components/LazyLoader';
-export { Tabs, TabsProvider, TabsConsumer } from './components/Tabs';
-export { StatusBadge, StatusBadgeGroup } from './components/StatusBadge';
-export { default as RangeSlider } from './components/RangeSlider';
-export { default as Typography } from './components/Typography';
+} from "./components/DeviceConnection";
+export { default as SeatTooltip } from "./components/Tooltip/SeatTooltip";
+export { default as SeatData } from "./components/Tooltip/SeatData";
+export { LazyLoader } from "./components/LazyLoader";
+export { Tabs, TabsProvider, TabsConsumer } from "./components/Tabs";
+export { StatusBadge, StatusBadgeGroup } from "./components/StatusBadge";
+export { default as RangeSlider } from "./components/RangeSlider";
+export { default as Typography } from "./components/Typography";
 export {
   fontSize,
   fontSizes,
@@ -241,5 +237,5 @@ export {
   textTransforms,
   themeColor,
   themeColors
-} from './mixins';
-export { default as getThemeValue } from './utils/getThemeValue';
+} from "./mixins";
+export { default as getThemeValue } from "./utils/getThemeValue";
